Final - Nov. 18, 2019 - Minor update

+++ src/MyFileWrite.java
@@ -1,4 +1,7 @@
 import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -16,37 +19,38 @@ public class MyFileWrite {
     public MyFileWrite() throws IOException {
         setFile();
         initializeObjects();
-    }
-
-    private void initializeObjects() throws IOException {
-        myPW = new PrintWriter(new BufferedWriter(new FileWriter(newFileName)));
-        time = new SimpleDateFormat("MM/dd/yyyy hh:mm:ss a");
-
         myPW.println(time.format(new Date(System.currentTimeMillis())));
     }
 
     private void setFile() {
         newFile = new File(newFileName);
         try {
-            if (newFile.exists()) {
-                newFile.delete();
-                System.out.println("[!] File exist");
-            } else {
-                System.out.println("[!] File not exist");
-            }
-            newFile.createNewFile();
+            if (newFile.exists()) newFile.delete();
+            if (newFile.createNewFile()) System.out.println("[!] Log file created");
         } catch (IOException e) {
             System.out.println("[!] Log module is offline");
         }
+    }
+
+    private void initializeObjects() throws IOException {
+        myPW = new PrintWriter(new BufferedWriter(new FileWriter(newFileName)));
+        time = new SimpleDateFormat("MM/dd/yyyy_hh:mm:ss");
     }
 
     public void recordMove(String move) {
         myPW.println((counter++) + ". " + move);
     }
 
-    public void endRecord(){
-        newFile.renameTo(new File("./log/" + time.format(new Date())));
+    public void endRecord() {
         myPW.close();
+        String tempString = time.format(new Date()) + ".txt";
+
+        File newFile = new File(tempString);
+        try {
+            Files.move(newFile.toPath(), newFile.toPath());
+        } catch (IOException e) {
+            System.out.println(e);
+        }
     }
